tentative fix for the size of this array while I wait for an answer from sai

+++ ecohydrology/cellular_automaton_vegetation_flat_surface/Ecohyd_functions_flat.py
@@ -111,7 +111,7 @@ def Empty_arrays(n, grid, grid1):
     Time = np.empty(n)  # To record time elapsed from the start of simulation
 #    CumWaterStress = np.empty([n/55, grid1.number_of_cells])
     # Cum Water Stress
-    VegType = np.empty([n/55, grid1.number_of_cells], dtype=int)
+    VegType = np.empty([int(n/55), grid1.number_of_cells], dtype=int)
     PET_ = np.zeros([365, grid.number_of_cells])
     Rad_Factor = np.empty([365, grid.number_of_cells])
     EP30 = np.empty([365, grid.number_of_cells])
